bot: renamed upload.py file and remove the defaultissue class there
